+++ MyFinancesApp/MyFinances/Expenses Feature/ExpenseItem.swift
@@ -10,10 +10,10 @@ import Foundation
 public struct ExpenseItem: Equatable {
     public let id: UUID
     public let title: String
-    public let amount: Float
+    public let amount: Double
     public let createdAt: Date
 
-    public init (id: UUID, title: String, amount: Float, createdAt: Date) {
+    public init (id: UUID, title: String, amount: Double, createdAt: Date) {
         self.id = id
         self.title = title
         self.amount = amount
